+++ testdata/gents_map_directive_bundle/input/b_user.go
@@ -1,0 +1,13 @@
+package api
+
+import (
+	"time"
+
+	"github.com/google/uuid"
+)
+
+//gents:export
+type User struct {
+	ID      uuid.UUID `json:"id"`
+	Created time.Time `json:"created"`
+}