+++ testdata/panic_gents_map_conflict/input/a.go
@@ -1,0 +1,3 @@
+//gents:map uuid.UUID=string
+
+package x