+++ testdata/panic_gents_map_conflict/input/b.go
@@ -1,0 +1,10 @@
+//gents:map uuid.UUID=number
+
+package x
+
+import "github.com/google/uuid"
+
+//gents:export
+type Foo struct {
+	ID uuid.UUID `json:"id"`
+}