+++ testdata/panic_gents_map_malformed/input.go
@@ -1,0 +1,8 @@
+//gents:map notavalidspec
+
+package x
+
+//gents:export
+type Foo struct {
+	Y int `json:"y"`
+}